update UI, login, role user, edit page

+++ SFCS/src/controller/admin/ProductAddController.java
@@ -56,7 +56,7 @@ public class ProductAddController extends HttpServlet {
 				} else if (item.getFieldName().equals("category")) {
 					product.setCategory(categoryService.get(Integer.parseInt(item.getString())));
 				} else if (item.getFieldName().equals("price")) {
-					product.setPrice(Long.parseLong(item.getString()));
+					product.setPrice(Integer.parseInt(item.getString()));
 				} else if (item.getFieldName().equals("des")) {
 					product.setDes(item.getString());;
 				} else if (item.getFieldName().equals("image")) {
